refactor(orchestrator): organize curriculum components
