+++ src/prime_rl/orchestrator/curriculum/__init__.py
@@ -1,13 +1,8 @@
 """Task selection and finalized-sample admission for training environments."""
 
-from prime_rl.orchestrator.curriculum.adv_range_gate import AdvRangeGate
-from prime_rl.orchestrator.curriculum.base import (
-    AdmissionGate,
-    Curriculum,
-    TaskSampler,
-)
-from prime_rl.orchestrator.curriculum.difficulty_pool_sampler import DifficultyPoolSampler
-from prime_rl.orchestrator.curriculum.standard_sampler import StandardSampler
+from prime_rl.orchestrator.curriculum.base import Curriculum
+from prime_rl.orchestrator.curriculum.gates import AdmissionGate, AdvRangeGate
+from prime_rl.orchestrator.curriculum.samplers import DifficultyPoolSampler, StandardSampler, TaskSampler
 
 __all__ = [
     "AdmissionGate",
