+++ src/prime_rl/orchestrator/curriculum/samplers/__init__.py
@@ -1,0 +1,7 @@
+"""Task sampler interfaces and implementations."""
+
+from prime_rl.orchestrator.curriculum.samplers.base import TaskSampler
+from prime_rl.orchestrator.curriculum.samplers.pool import DifficultyPoolSampler
+from prime_rl.orchestrator.curriculum.samplers.simple import StandardSampler
+
+__all__ = ["DifficultyPoolSampler", "StandardSampler", "TaskSampler"]
